+++ src/generate/gen_propsheet_dlg.cpp
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////
 // Purpose:   wxPropertySheetDialog generator
 // Author:    Ralph Walden
-// Copyright: Copyright (c) 2023 KeyWorks Software (Ralph Walden)
+// Copyright: Copyright (c) 2023-2025 KeyWorks Software (Ralph Walden)
 // License:   Apache License -- see ../../LICENSE
 /////////////////////////////////////////////////////////////////////////////
 
@@ -153,8 +153,9 @@ bool PropSheetDlgGenerator::ConstructionCode(Code& code)
     }
     else if (code.is_ruby())
     {
-        code.Add("class ").NodeName().Add(" < Wx::PropertySheetDialog");
-        code.Eol().Tab().Add("def initialize(parent");
+        code.Add("class ").NodeName().Add(" < Wx::PropertySheetDialog").Eol();
+        code.AddPublicRubyMembers();
+        code.Eol(eol_if_needed).Tab().Add("def initialize(parent");
         // Indent any wrapped lines
         code.Indent(3);
         code.Str(", id = ");